pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "option-chain-dashboard"
version = "0.1.0"
description = "A comprehensive dashboard for analyzing options chains, volatility, and trading opportunities"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Dashboard Team", email = "dashboard@example.com"}
]
keywords = ["options", "trading", "dashboard", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "streamlit==1.28.1",
    "yfinance==0.2.32",
    "pandas~=2.1.0",
    "numpy~=1.24.0",
    "duckdb==0.9.2",
    "scipy~=1.11.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "pytz==2023.3.post1",
    "requests==2.31.0",
    "pyyaml==6.0.1",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.21.1",
]
dev = [
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
    "isort==5.13.2",
]

[project.urls]
Homepage = "https://github.com/example/option-chain-dashboard"
Documentation = "https://option-chain-dashboard.readthedocs.io"
Repository = "https://github.com/example/option-chain-dashboard.git"
Issues = "https://github.com/example/option-chain-dashboard/issues"

[tool.setuptools]
packages = ["functions", "ui", "scripts"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".pytest_cache",
    ".mypy_cache",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C901",   # mccabe complexity
    "B",      # flake8-bugbear
]
ignore = [
    "E501",   # line too long (handled by black)
    "W605",   # invalid escape sequence (handled by black)
]

[tool.ruff.lint.isort]
known-first-party = ["functions", "ui", "scripts"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
